Allow different solution types

+++ src/day_07.rs
@@ -20,26 +20,30 @@ fn median(input: &Vec<u32>) -> u32 {
 
 /* Calculate fuel consumption */
 
-fn get_steps(from: u32, to: u32) -> u64 {
-    (if from > to { from - to } else { to - from }) as u64
+fn get_steps(from: u32, to: u32) -> u32 {
+    if from > to {
+        from - to
+    } else {
+        to - from
+    }
 }
 
-fn calc_trip(from: u32, to: u32) -> u64 {
+fn calc_trip(from: u32, to: u32) -> u32 {
     let steps = get_steps(from, to);
     (steps * (steps + 1)) / 2
 }
 
-fn calc_fuel<F: Fn(u32, u32) -> u64>(calculator: F, positions: &Vec<u32>, position: u32) -> u64 {
+fn calc_fuel<F: Fn(u32, u32) -> u32>(calculator: F, positions: &Vec<u32>, position: u32) -> u32 {
     positions.iter().map(|&x| calculator(position, x)).sum()
 }
 
 /* Solutions */
 
-fn part01(input: &Vec<u32>) -> u64 {
+fn part01(input: &Vec<u32>) -> u32 {
     calc_fuel(get_steps, input, median(input))
 }
 
-fn part02(input: &Vec<u32>) -> u64 {
+fn part02(input: &Vec<u32>) -> u32 {
     let fuel_cost = means(input).map(|x| calc_fuel(calc_trip, input, x));
     fuel_cost[0].min(fuel_cost[1])
 }
@@ -53,22 +57,22 @@ pub fn day_07() -> Solution {
 /* Brute force solution - for fun */
 
 #[allow(dead_code)]
-fn brute<F: Fn(u32, u32) -> u64 + Copy>(calc: F, input: &Vec<u32>) -> u64 {
+fn brute<F: Fn(u32, u32) -> u32 + Copy>(calc: F, input: &Vec<u32>) -> u32 {
     let min = *input.iter().min().unwrap();
     let max = *input.iter().max().unwrap() + 1;
 
     (min..max)
         .into_iter()
-        .fold(u64::MAX, |cost, x| cost.min(calc_fuel(calc, &input, x)))
+        .fold(u32::MAX, |cost, x| cost.min(calc_fuel(calc, &input, x)))
 }
 
 #[allow(dead_code)]
-fn brute_01(input: &Vec<u32>) -> u64 {
+fn brute_01(input: &Vec<u32>) -> u32 {
     brute(get_steps, input)
 }
 
 #[allow(dead_code)]
-fn brute_02(input: &Vec<u32>) -> u64 {
+fn brute_02(input: &Vec<u32>) -> u32 {
     brute(calc_trip, input)
 }
